Migrate 18 experiments to PersistableExperiment + add Axis.dtype (ADR-0027 P-C)
Switch 18 in-scope single-file experiments (onetone, twotone freq/power/flux/
ac_stark/zigzag/rabi/t1/t2echo/t2ramsey, jpa, lookback; 21 classes) from the
dict save_data/load_data API to native-labber persistence via a per-experiment
AXES_SPEC on PersistableExperiment; delete their hand-written save/load. Add an
optional Axis.dtype (default float64) so int64 sweep axes (zigzag counts) restore
their dtype on load. Multi-class files get one AXES_SPEC each; tuple-returning
run()s (t2echo/t2ramsey) keep a manual last_result instead of @record_result.
pyright experiment/v2: 0 errors.

+++ lib/zcu_tools/experiment/base.py
@@ -217,7 +217,9 @@ class PersistableExperiment(AbsExperiment[T_Result, T_Config]):
                 cfg_snapshot = spec.cfg_type.validate_or_warn(cfg_dict, source=filepath)
 
         kwargs: dict[str, Any] = {
-            ax.field_name: np.asarray(ld.axes[i].values, dtype=np.float64) / ax.scale
+            ax.field_name: (
+                np.asarray(ld.axes[i].values, dtype=np.float64) / ax.scale
+            ).astype(ax.dtype)
             for i, ax in enumerate(spec.axes)
         }
         kwargs[spec.z.field_name] = np.asarray(ld.z).astype(spec.z.dtype)
